Get Coaching History and Notable Recruits

Returns coaching history with four factor and tournament finish (tibble one) along with notable recruits (tibble two).

Usage

bart_coach(coach)

Arguments

coach

Coach to return.

Value

Returns a list of tibbles: 'Coach History' and 'Recruiting History'

Details

For a brief explanation of each factor and its computation, please visit KenPom's blog. `rec_rank` is a scale of recruiting ranking, where 100 is a consensus #1 recruit and 0 is unrecruited.

Examples

bart_coach('Mike Krzyzewski')
#> [[1]]
#> # A tibble: 15 × 23
#>     year  t_rk team   seed finish   conf  ov_rec conf_…¹ adj_o adj_d efg_o efg_d
#>    <dbl> <dbl> <chr> <dbl> <chr>    <chr> <chr>  <chr>   <dbl> <dbl> <dbl> <dbl>
#>  1  2022     6 Duke      2 Final F… ACC   32-7   16-4     123.  96.1  55.6  47.1
#>  2  2021    32 Duke     NA NA       ACC   13-11  9-9      115.  97.2  53.6  51.3
#>  3  2020     6 Duke      3 NA       ACC   25-6   15-5     115.  91.9  52.6  45.7
#>  4  2019     5 Duke      1 Elite E… ACC   32-6   14-4     120.  90.5  53.7  45.1
#>  5  2018     3 Duke      2 Elite E… ACC   29-8   13-5     122.  93.1  55.9  46.4
#>  6  2017    10 Duke      2 R32      ACC   28-9   11-7     122.  95.5  54.8  47.4
#>  7  2016    20 Duke      4 Sweet S… ACC   25-11  11-7     121.  98.9  53.7  49.5
#>  8  2015     2 Duke      1 CHAMPS   ACC   35-4   15-3     125.  90.5  56.6  46.6
#>  9  2014     6 Duke      3 R64      ACC   26-9   13-5     126.  98.7  53.8  49.3
#> 10  2013     8 Duke      2 Elite E… ACC   30-6   14-4     118.  91.5  53.9  45.6
#> 11  2012    18 Duke      2 R64      ACC   27-7   13-3     116.  95.4  52.8  47.2
#> 12  2011     3 Duke      1 Sweet S… ACC   32-5   13-3     120.  90.0  53.7  44.5
#> 13  2010     1 Duke      1 CHAMPS   ACC   35-5   13-3     121.  87.9  50.6  43.8
#> 14  2009    12 Duke      2 Sweet S… ACC   30-7   11-5     117.  92.2  50.5  47.8
#> 15  2008     7 Duke      2 R32      ACC   28-6   13-3     116.  89.9  53.6  47.7
#> # … with 11 more variables: to_percent <dbl>, tod_percent <dbl>,
#> #   or_percent <dbl>, dr_percent <dbl>, ftr <dbl>, ftrd <dbl>,
#> #   ft_percent <dbl>, two_pct <dbl>, two_pct_d <dbl>, three_pct <dbl>,
#> #   three_pct_d <chr>, and abbreviated variable name ¹​conf_rec
#> # ℹ Use `colnames()` to see all variable names
#> 
#> [[2]]
#> # A tibble: 63 × 4
#>     year player            team  rec_rank
#>    <int> <chr>             <chr>    <dbl>
#>  1  2018 Marvin Bagley III Duke     100  
#>  2  2019 R.J. Barrett      Duke     100  
#>  3  2015 Jahlil Okafor     Duke     100  
#>  4  2022 Paolo Banchero    Duke      99.8
#>  5  2017 Harry Giles       Duke      99.8
#>  6  2019 Cam Reddish       Duke      99.8
#>  7  2012 Austin Rivers     Duke      99.8
#>  8  2011 Kyrie Irving      Duke      99.8
#>  9  2016 Brandon Ingram    Duke      99.6
#> 10  2014 Jabari Parker     Duke      99.4
#> # … with 53 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#>