Get Conference Four Factor Statistics

Returns conference-wide four factor data on a variety of splits, including date range, quadrant level, opponent ranking, game location, and game type.

Usage

bart_conf_factors(
  year = current_season(),
  venue = "all",
  type = "all",
  quad = "4",
  top = 0,
  start = NULL,
  end = NULL
)

Arguments

year

Defaults to current season (YYYY).

venue

Filters by venue; defaults to `all`.

type

Filters by game type; defaults to `all`.

quad

Filters by quadrant level; defaults to `4`.

top

Filters by opponent T-Rank position; defaults to NULL (all).

start

Filters by start date; defaults to NULL (full season).

end

Filters by end date; defaults to NULL (full season).

Value

Returns a tibble with 22 columns:

conf

character.

barthag

double. The estimation of a team's win probability against the average Division 1 team on a neutral court.

rec

character.

wins

double.

games

double.

adj_t

double.

adj_o

double.

off_efg

double.

off_to

double.

off_or

double.

off_ftr

double.

adj_d

double.

def_efg

double.

def_to

double.

def_or

double.

def_ftr

double.

wab

double. The number of wins above or below the expected total from a bubble team against the same schedule.

year

double.

venue

character. Split supplied to the venue argument.

type

character. Split supplied to the type argument.

top

double. Split supplied to the top argument.

quad

character. Split supplied to the quad argument.

Details

For a brief explanation of each factor and its computation, please visit KenPom's blog. Data can be split on five variables:

venue

Splits on game location; 'all', 'home', 'away', 'neutral', and 'road' (away + neutral).

type

Splits on game type; 'all', 'nc' (non-conference), 'conf' (conference), 'reg' (regular season), 'post' (post-season tournaments), 'ncaa' (NCAA tournament).

quad

Splits by quadrant level; 1-4 with 0 indicating 1-A games.

top

Splits by opponent T-Rank position, adjusted for game location.

start/end

Splits by date range (YYYYMMDD).

Examples

bart_conf_factors(type='nc')
#> # A tibble: 32 × 22
#>    conf  barthag rec     wins games adj_t adj_o off_efg off_to off_or off_ftr
#>    <chr>   <dbl> <chr>  <dbl> <dbl> <dbl> <dbl>   <dbl>  <dbl>  <dbl>   <dbl>
#>  1 B12     0.881 105–22   105   127  68.1  107.    53.1   18.8   33.9    31.4
#>  2 B10     0.834 108–31   108   139  69    109.    53.5   17.6   30.9    33.6
#>  3 SEC     0.820 129–47   129   176  69.8  106.    50.8   18.1   33.8    31.1
#>  4 BE      0.791 88–27     88   115  70.2  107.    51.7   18.5   32.2    33.4
#>  5 P12     0.748 82–41     82   123  69.2  105.    50.6   17.6   31.7    32.1
#>  6 ACC     0.746 106–53   106   159  68.5  106.    51.8   17.4   29.5    32.5
#>  7 Amer    0.718 82–42     82   124  68.3  104.    50.2   18.1   31.7    32  
#>  8 MWC     0.698 83–41     83   124  68    104.    51.6   17.9   26.5    31  
#>  9 WCC     0.675 81–51     81   132  69.4  103.    51.9   19.4   28.8    29.2
#> 10 A10     0.609 98–66     98   164  68.4  103.    51.4   18.4   28.5    30.9
#> # … with 22 more rows, and 11 more variables: adj_d <dbl>, def_efg <dbl>,
#> #   def_to <dbl>, def_or <dbl>, def_ftr <dbl>, wab <dbl>, year <dbl>,
#> #   venue <chr>, type <chr>, top <dbl>, quad <chr>
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names