Get Conference Team Statistics

Returns conference-only metrics, strengths of schedule, and bid probabilities (auto and at-large) for all teams in a conference.

Usage

bart_conf_stats(year = current_season(), conf = NULL)

Arguments

year

Defaults to current season (YYYY).

conf

Indicates conference (see details).

Details

Accepted conference abbreviations for the `conf` argument are:

‘A10’, ‘ACC’, ‘AE’, ‘ASun’, ‘Amer’, ‘B10’, ‘B12’, ‘BE’, ‘BSky’, ‘BSth’, ‘BW’, ‘CAA’, ‘CUSA’, ‘Horz’, ‘Ivy’, ‘MAAC’, ‘MAC’, ‘MEAC’, ‘MVC’, ‘MWC’, ‘NEC’, ‘OVC’, ‘P12’, ‘Pat’, ‘SB’, ‘SC’, ‘SEC’, ‘SWAC’, ‘Slnd’, ‘Sum’, ‘WAC’, ‘WCC’

Other:

`conf_cur_sos` is the current average Barthag rating of conference opponents.

`conf_fut_sos` is the projected average Barthag rating of conference opponents.

Examples

bart_conf_stats(year=2022, conf='ACC')
#> # A tibble: 15 × 23
#>       rk team         seed finish conf_rec adj_oe adj_de barthag eff_marg con_oe
#>    <dbl> <chr>       <dbl> <chr>  <chr>     <dbl>  <dbl>   <dbl>    <dbl>  <dbl>
#>  1     6 Duke            2 Final… 16-4      123.    96.1   0.944       28   121.
#>  2    16 North Caro…     8 Finals 15-5      114.    93.5   0.911       18   115 
#>  3    36 Notre Dame     11 R32    15-5      113.    96.8   0.854       18   114.
#>  4    38 Miami FL       10 Elite… 14-6      114.    98.0   0.847       17   117.
#>  5    35 Wake Forest    NA NA     13-7      111.    95.0   0.854       20   115.
#>  6    59 Virginia       NA NA     12-8      107.    95.1   0.801       14   112.
#>  7    18 Virginia T…    11 R64    11-9      116.    95.2   0.903       16   115.
#>  8   103 Florida St.    NA NA     10-10     105.    98.4   0.683        7   106 
#>  9    45 Syracuse       NA NA     9-11      116.   101.    0.827       13   115.
#> 10    70 Clemson        NA NA     8-12      109.    98.4   0.775       10   109 
#> 11   111 Boston Col…    NA NA     6-14      104.    98.8   0.642        4   103 
#> 12   118 Louisville     NA NA     6-14      103.    98.3   0.631        1   104 
#> 13   195 Pittsburgh     NA NA     6-14       98.9  101.    0.438        0   103.
#> 14   139 Georgia Te…    NA NA     5-15      101.    97.9   0.579        4   101.
#> 15   120 North Caro…    NA NA     4-16      109.   105.    0.625        7   112.
#> # … with 13 more variables: con_oe_rk <dbl>, con_de <dbl>, con_de_rk <dbl>,
#> #   conf_barthag <dbl>, proj_rec <chr>, conf_cur_sos <dbl>,
#> #   conf_cur_sos_rk <dbl>, conf_fut_sos <dbl>, conf_fut_sos_rk <dbl>,
#> #   conf_sos <dbl>, conf_sos_rk <dbl>, auto_prob <dbl>, bid_prob <dbl>