Get Four Factor Statistics

Returns four factor data and team records on a variety of splits, including date range, quadrant level, opponent ranking, game location, and game type.

Usage

bart_factors(
  year = current_season(),
  venue = "all",
  type = "all",
  quad = "4",
  top = 0,
  start = NULL,
  end = NULL
)

Arguments

year

Defaults to current season (YYYY).

venue

Filters by venue; defaults to `all`.

type

Filters by game type; defaults to `all`.

quad

Filters by quadrant level; defaults to `4`.

top

Filters by opponent T-Rank position; defaults to NULL (all).

start

Filters by start date; defaults to NULL (full season).

end

Filters by end date; defaults to NULL (full season).

Value

Returns a tibble with 22 columns:

team

character.

conf

character.

barthag

double. The estimation of a team's win probability against the average Division 1 team on a neutral court.

rec

character.

wins

double.

games

double.

adj_t

double.

adj_o

double.

off_efg

double.

off_to

double.

off_or

double.

off_ftr

double.

adj_d

double.

def_efg

double.

def_to

double.

def_or

double.

def_ftr

double.

wab

double. The number of wins above or below the expected total from a bubble team against the same schedule.

year

double.

venue

character. Split supplied to the venue argument.

type

character. Split supplied to the type argument.

top

double. Split supplied to the top argument.

quad

character. Split supplied to the quad argument.

Details

For a brief explanation of each factor and its computation, please visit KenPom's blog. Data can be split on five variables:

venue

Splits on game location; 'all', 'home', 'away', 'neutral', and 'road' (away + neutral).

type

Splits on game type; 'all', 'nc' (non-conference), 'conf' (conference), 'reg' (regular season), 'post' (post-season tournaments), 'ncaa' (NCAA tournament).

quad

Splits by quadrant level; 1-4 with 0 indicating 1-A games.

top

Splits by opponent T-Rank position, adjusted for game location.

start/end

Splits by date range (YYYYMMDD).

Examples

bart_factors(quad='3', venue='away', start='20220101')
#> # A tibble: 357 × 24
#>    team       conf  barthag rec    wins games adj_t adj_o off_efg off_to off_or
#>    <chr>      <chr>   <dbl> <chr> <dbl> <dbl> <dbl> <dbl>   <dbl>  <dbl>  <dbl>
#>  1 Gonzaga    WCC     0.981 4–1       4     5  78.3  118.    58.5   16.5   20  
#>  2 Duke       ACC     0.979 9–1       9    10  65.5  126.    56.6   14.5   32.6
#>  3 Arizona    P12     0.960 7–2       7     9  73.8  122.    55.9   18.5   33.7
#>  4 Kentucky   SEC     0.948 6–4       6    10  67.6  116.    49.4   18.2   32.8
#>  5 Murray St. OVC     0.943 2–0       2     2  64.1  112.    50.9   14.9   28.6
#>  6 Vermont    AE      0.938 3–0       3     3  66.1  116.    66.1   16.4   21.3
#>  7 Kansas     B12     0.930 5–4       5     9  68.1  120.    51.1   20.3   35.1
#>  8 Baylor     B12     0.929 7–3       7    10  68.4  116.    50.6   20.3   32.9
#>  9 Iowa       B10     0.918 5–4       5     9  69.1  121.    52.3   12.7   31.1
#> 10 Villanova  BE      0.913 7–2       7     9  66.2  120.    54.1   18.7   30.1
#> # … with 347 more rows, and 13 more variables: off_ftr <dbl>, adj_d <dbl>,
#> #   def_efg <dbl>, def_to <dbl>, def_or <dbl>, def_ftr <dbl>, wab <dbl>,
#> #   year <dbl>, venue <chr>, type <chr>, top <dbl>, quad <chr>, start <chr>