Get Team Game Stats

Returns box score data by team for all games.

Usage

bart_game_box(year = current_season())

Arguments

year

Defaults to current season (YYYY).

Value

Returns a tibble of box score statistics

Details

The home team is coded as `team2`. Neutral site games may contain errors as to whom is the home team.

Examples

bart_game_box(year=2022)
#> # A tibble: 5,967 × 38
#>    date         min team1          team2 team1_fgm team1_fga team1_tpm team1_tpa
#>    <date>     <dbl> <chr>          <chr>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1 2022-04-04   200 North Carolina Kans…        23        73         5        23
#>  2 2022-04-02   200 Villanova      Kans…        22        57        13        31
#>  3 2022-04-02   200 North Carolina Duke         27        64        10        26
#>  4 2022-04-01   200 Fresno St.     Coas…        31        59        12        20
#>  5 2022-03-31   200 Xavier         Texa…        29        60         7        18
#>  6 2022-03-29   200 St. Bonaventu… Xavi…        29        62         9        22
#>  7 2022-03-29   200 Washington St. Texa…        19        55         7        29
#>  8 2022-03-28   225 Coastal Carol… Sout…        28        59         9        24
#>  9 2022-03-28   200 Southern Utah  Fres…        19        61         4        20
#> 10 2022-03-27   200 Miami FL       Kans…        19        55         3        21
#> # … with 5,957 more rows, and 30 more variables: team1_ftm <dbl>,
#> #   team1_fta <dbl>, team1_oreb <dbl>, team1_dreb <dbl>, team1_reb <dbl>,
#> #   team1_ast <dbl>, team1_stl <dbl>, team1_blk <dbl>, team1_to <dbl>,
#> #   team1_pf <dbl>, team1_pts <dbl>, team2_fgm <dbl>, team2_fga <dbl>,
#> #   team2_tpm <dbl>, team2_tpa <dbl>, team2_ftm <dbl>, team2_fta <dbl>,
#> #   team2_oreb <dbl>, team2_dreb <dbl>, team2_reb <dbl>, team2_ast <dbl>,
#> #   team2_stl <dbl>, team2_blk <dbl>, team2_to <dbl>, team2_pf <dbl>, …