Get Game Four Factors

Returns game-by-game four factor statistics.

Usage

bart_game_factors(year = current_season())

Arguments

year

Defaults to current season (YYYY).

Value

Returns a tibble of four factor statistics

Details

For a brief explanation of each factor and its computation, please visit KenPom's blog. `avg_marg` and `opp_avg_marg` is the the average lead or deficit during a game.

Examples

bart_game_factors(year=2022)
#> # A tibble: 10,950 × 28
#>    date       type  team       conf  opp   opp_c…¹ loc   result avg_marg opp_a…²
#>    <date>     <chr> <chr>      <chr> <chr> <chr>   <chr> <chr>     <dbl>   <dbl>
#>  1 2022-04-04 post  Kansas     B12   Nort… ACC     N     W, 72…  -2.09     0.911
#>  2 2022-04-04 post  North Car… ACC   Kans… B12     N     L, 72…   2.09     0.958
#>  3 2022-04-02 post  Duke       ACC   Nort… ACC     N     L, 81…   0.0621   0.911
#>  4 2022-04-02 post  North Car… ACC   Duke  ACC     N     W, 81…  -0.0621   0.944
#>  5 2022-04-02 post  Kansas     B12   Vill… BE      N     W, 81…  10.9      0.935
#>  6 2022-04-02 post  Villanova  BE    Kans… B12     N     L, 81… -10.9      0.958
#>  7 2022-04-01 post  Coastal C… SB    Fres… MWC     H     L, 85… -13.2      0.739
#>  8 2022-04-01 post  Fresno St. MWC   Coas… SB      A     W, 85…  13.2      0.589
#>  9 2022-03-31 post  Texas A&M  SEC   Xavi… BE      N     L, 73…  -0.0246   0.821
#> 10 2022-03-31 post  Xavier     BE    Texa… SEC     N     W, 73…   0.0246   0.860
#> # … with 10,940 more rows, 18 more variables: adj_o <dbl>, adj_d <dbl>,
#> #   off_ppp <dbl>, off_efg <dbl>, off_to <dbl>, off_or <dbl>, off_ftr <dbl>,
#> #   def_ppp <dbl>, def_efg <dbl>, def_to <dbl>, def_or <dbl>, def_ftr <dbl>,
#> #   game_score <dbl>, season <dbl>, tempo <dbl>, game_id <chr>, coach <chr>,
#> #   opp_coach <chr>, and abbreviated variable names ¹​opp_conf, ²​opp_avg_marg
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names