Get Player Season Splits

Returns season-long player statistics on a variety of splits.

Usage

bart_player_splits(
  year = current_season(),
  split = NULL,
  team = NULL,
  conf = NULL,
  exp = NULL,
  player_id = NULL,
  type = NULL
)

Arguments

year

Year to filter.

split

Split to filter (see details).

team

Team to filter.

conf

Conference to filter.

exp

Experience to filter.

player_id

Unique player ID to filter.

type

Data type to return (see details).

Value

Returns a tibble with the number of columns dependent on the value supplied to the `stat` argument.

Details

Data is split on three statistical types, explained below:

box

Returns basic box score stats; sorts by ppg.

shooting

Returns play-by-play shooting splits; sorts by ppg.

advanced

Returns advanced metrics and possession-adjusted box score statistics; sorts by recruiting rank.

Examples

bart_player_splits(year=2022, conf='ACC')
#> ── Player Splits ─────────────────────────────────────────────── toRvik 1.1.0 ──
#> ℹ Data updated: 2022-09-13 04:53:43 UTC
#> # A tibble: 411 × 44
#>    player    id team  result   min   pts two_m two_a three_m three_a   ftm   fta
#>    <chr>  <dbl> <chr> <chr>  <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl> <dbl>
#>  1 AJ Gr… 75269 Duke  L       24.4  5.86  1.29  2.86    0.86    3.71  0.71  1   
#>  2 AJ Gr… 75269 Duke  W       24.3 11.4   2.06  3.66    2.03    4.16  1.16  1.44
#>  3 Abe A… 72612 Bost… W        1    0.33  0     0       0       0     0.33  0.67
#>  4 Aidan… 72942 Pitt… L        1    0     0     0.5     0       0     0     0   
#>  5 Aidan… 72942 Pitt… W        0.5  0     0     0       0       0     0     0   
#>  6 Al-Am… 70627 Clem… L       29.6  9.75  0.69  2       2.5     6.19  0.88  1.12
#>  7 Al-Am… 70627 Clem… W       29.8 12.7   1.41  3       2.59    6.59  2.12  2.41
#>  8 Alex … 70621 Clem… L       16.1  4.79  0.43  0.64    1.07    3.57  0.71  1   
#>  9 Alex … 70621 Clem… W       13.4  5.53  0.29  0.94    1.41    2.71  0.71  1   
#> 10 Alex … 75787 Nort… L        1    0     0     0       0       0     0     0   
#> # … with 401 more rows, and 32 more variables: oreb <dbl>, dreb <dbl>,
#> #   ast <dbl>, tov <dbl>, stl <dbl>, blk <dbl>, pf <dbl>, ortg <dbl>,
#> #   usg <dbl>, fg_pct <dbl>, efg <dbl>, ts <dbl>, dunk_m <dbl>, dunk_a <dbl>,
#> #   rim_m <dbl>, rim_a <dbl>, mid_m <dbl>, mid_a <dbl>, or_pct <dbl>,
#> #   dr_pct <dbl>, ast_pct <dbl>, to_pct <dbl>, stl_pct <dbl>, blk_pct <dbl>,
#> #   bpm <dbl>, obpm <dbl>, dbpm <dbl>, net <dbl>, poss <dbl>, year <dbl>,
#> #   games <int>, conf <chr>