Get Pregame Probabilities

Returns pregame win probabilities, expected scores, and thrill quotients for all games.

Usage

bart_pregame(year = current_season(), team = NULL, type = NULL, game_id = NULL)

Arguments

year

Defaults to current season (YYYY).

team

Team to filter for.

type

Game type to filter for ('nc', 'conf', 'conf_t', 'post', 'nond1')

game_id

Game ID to filter for.

Value

Returns a tibble with 13 columns:

date

double.

conf

character.

line

character.

ttq

double. Torvik Thrill Quotient -- measures how good the teams are, how close the game is projected to be, and how fast the tempo is projected to be.

type

character. See details.

team1

character.

team1_wp

double. Estimated win percentage.

team1_pts

double. Estimated total points.

team2

character.

team2_wp

double. Estimated win percentage.

team2_pts

double. Estimated total points.

game_id

character.

year

double.

Details

Under the `type` column, games are classified by one of five indicators, explained below:

nc

Non-conference games between two D-1 teams.

conf

In-conference games.

conf_t

Conference tournament games.

post

Post-conference tournament games.

nond1

Games involving one non-D1 team.

Examples

bart_pregame(year=2022)
#> ── Pregame Probabilities ─────────────────────────────────────── toRvik 1.1.0 ──
#> ℹ Data updated: 2022-09-09 12:24:16 UTC
#> # A tibble: 5,967 × 13
#>    date      conf  line    ttq type  team1 team1…¹ team1…² team2 team2…³ team2…⁴
#>    <chr>     <chr> <chr> <dbl> <chr> <chr>   <dbl>   <dbl> <chr>   <dbl>   <dbl>
#>  1 2021-11-… ACC … Duke…  91.5 nc    Duke   0.561     77.9 Kent…   0.439    76.2
#>  2 2021-11-… B12 … Kans…  70.7 nc    Kans…  0.812     78.7 Mich…   0.188    69.3
#>  3 2021-11-… OVC … Belm…  66.9 nc    Belm…  0.531     71.0 Ohio    0.469    70.2
#>  4 2021-11-… WAC … Utah…  60.2 nc    Abil…  0.374     70.6 Utah    0.626    73.8
#>  5 2021-11-… MVC … Sout…  59.5 nc    Brad…  0.279     70.7 Sout…   0.721    76.8
#>  6 2021-11-… CUSA… Alab…  58.0 nc    Loui…  0.161     71.5 Alab…   0.839    82.5
#>  7 2021-11-… CAA … Illi…  55.4 nc    UNC …  0.445     69.9 Illi…   0.555    71.3
#>  8 2021-11-… MAC … Tole…  53.9 nc    Tole…  0.641     73.3 Valp…   0.359    69.7
#>  9 2021-11-… MAAC… VCU …  53.6 nc    Sain…  0.214     56.4 VCU     0.786    63.1
#> 10 2021-11-… CAA … Hous…  51.0 nc    Hofs…  0.0395    60.1 Hous…   0.960    79.4
#> # … with 5,957 more rows, 2 more variables: game_id <chr>, year <int>, and
#> #   abbreviated variable names ¹​team1_wp, ²​team1_pts, ³​team2_wp, ⁴​team2_pts