Get T-Rank Ratings

Returns current T-Rank ratings and two forms of strength of schedule.

Usage

bart_ratings(year = current_season())

Arguments

year

Defaults to current season (YYYY).

Value

Returns a tibble with 19 columns:

team

character.

conf

character.

barthag

double. The estimation of a team's win probability against the average Division 1 team on a neutral court.

barthag_rk

integer.

adj_o

double.

adj_o_rk

integer.

adj_d

double.

adj_d_rk

integer.

adj_t

double.

adj_t_rk

integer.

wab

double. The number of wins above or below the expected total from a bubble team against the same schedule.

nc_elite_sos

double.

nc_fut_sos

double.

nc_cur_sos

double.

ov_elite_sos

double.

ov_fut_sos

double.

ov_cur_sos

double.

seed

integer.

year

double.

Details

`x_cur_sos` is the current average Barthag rating of opponents.

`x_fut_sos` is the projected average Barthag rating of opponents.

Examples

bart_ratings(year=2020)
#> ── Team Ratings: 2020 ────────────────────────────────────────── toRvik 1.1.0 ──
#> ℹ Data updated: 2022-09-19 21:54:05 UTC
#> # A tibble: 353 × 19
#>    team    conf  barthag barth…¹ adj_o adj_o…² adj_d adj_d…³ adj_t adj_t…⁴   wab
#>    <chr>   <chr>   <dbl>   <int> <dbl>   <int> <dbl>   <int> <dbl>   <int> <dbl>
#>  1 Kansas  B12     0.962       1  116.       7  87.7       2  67.4     226 10.9 
#>  2 Baylor  B12     0.951       2  114.      13  88.4       3  66.2     273  8.54
#>  3 Gonzaga WCC     0.947       3  121.       1  94.3      41  72.0      33  7.69
#>  4 Dayton  A10     0.945       4  120.       3  93.4      29  67.5     219  6.81
#>  5 Michig… B10     0.933       5  115.      11  91.3      11  69.3     117  5.23
#>  6 Duke    ACC     0.931       6  115.       9  91.9      17  71.7      39  5.09
#>  7 Creigh… BE      0.929       7  121.       2  96.4      67  68.3     170  6.10
#>  8 Ohio S… B10     0.920       8  115.      12  92.6      22  66.2     278  3.80
#>  9 San Di… MWC     0.912       9  114.      17  92.8      25  64.7     330  6.64
#> 10 Louisv… ACC     0.912      10  115.      10  94.0      37  66.9     244  4.34
#> # … with 343 more rows, 8 more variables: nc_elite_sos <int>, nc_fut_sos <dbl>,
#> #   nc_cur_sos <dbl>, ov_elite_sos <int>, ov_fut_sos <dbl>, ov_cur_sos <dbl>,
#> #   seed <lgl>, year <int>, and abbreviated variable names ¹​barthag_rk,
#> #   ²​adj_o_rk, ³​adj_d_rk, ⁴​adj_t_rk