Get Full Season Schedule

Returns full season schedule for all D-1 teams by year

Usage

bart_season_schedule(year = current_season())

Arguments

year

Defaults to current season (YYYY).

Value

Returns a tibble with six columns:

date

double.

type

character. See details.

neutral

logical.

home

character.

away

character.

game_id

character.

Details

Function pulls all games involving at least one D-1 team. Under the `type` column, games are classified by one of four indicators, explained below:

nc

Non-conference games between two D-1 teams.

conf

In-conference games.

post

Post-conference tournament games.

nond1

Games involving one non-D1 team.

Examples

bart_season_schedule(year=2022)
#> # A tibble: 5,968 × 6
#>    date       type  neutral home               away                game_id      
#>    <date>     <chr> <lgl>   <chr>              <chr>               <chr>        
#>  1 2021-11-09 nc    FALSE   Creighton          Arkansas Pine Bluff Arkansas Pin…
#>  2 2021-11-09 nc    FALSE   Valparaiso         Toledo              ToledoValpar…
#>  3 2021-11-09 nc    FALSE   Indiana            Eastern Michigan    Eastern Mich…
#>  4 2021-11-09 nc    FALSE   North Carolina St. Bucknell            BucknellNort…
#>  5 2021-11-09 nc    FALSE   Missouri           Central Michigan    Central Mich…
#>  6 2021-11-09 nond1 FALSE   Brown              Salve Regina        Salve Regina…
#>  7 2021-11-09 nond1 FALSE   Wright St.         Lake Erie           Lake ErieWri…
#>  8 2021-11-09 nc    FALSE   New Mexico St.     UC Irvine           UC IrvineNew…
#>  9 2021-11-09 nond1 FALSE   Furman             North Greenville    North Greenv…
#> 10 2021-11-09 nc    FALSE   UAB                UNC Asheville       UNC Ashevill…
#> # … with 5,958 more rows