Get Team Box Stats

Returns team box totals and per-game averages by game type back to 2008.

Usage

bart_team_box(year = current_season(), type = "all")

Arguments

year

Defaults to current season (YYYY).

type

Filters by game type; defaults to `all`.

Value

Returns a tibble of team box totals and per-game averages

Details

Columns ending in 'pg' indicate a per-game average. Unless specified by `type=d1`, results include games against non-Division 1 opponents.

The `type` argument splits the results by game type, explained below:

all

All games played.

nc

Non-conference games.

conf

In-conference games.

post

Post-conference tournament games.

d1

Games against D-1 teams only.

nond1

Games involving one non-D1 team.

Examples

bart_team_box(type='conf')
#> # A tibble: 358 × 27
#>    team    ast   blk  dreb   fga   fgm   fta   ftm  oreb    pf   pts   reb   stl
#>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Gonz…   266    83   471   912   480   238   182   121   218  1271   592   101
#>  2 Sout…   267    49   518  1024   562   361   286   106   271  1568   624   101
#>  3 Oral…   214    77   570  1258   594   302   224   191   314  1619   761    85
#>  4 Tole…   325    65   600  1256   626   349   270   202   300  1669   802   108
#>  5 Sout…   200    49   336   795   383   360   279   115   242  1163   451   111
#>  6 Brya…   249    68   507  1076   483   397   298   213   295  1411   720   105
#>  7 Ariz…   386   101   588  1196   592   424   309   219   319  1641   807   121
#>  8 Nort…   250    54   543  1163   581   367   267   138   362  1629   681    88
#>  9 Hous…   182    27   334   831   379   368   274   149   289  1132   483   127
#> 10 New …   160    28   394   813   388   379   282   121   273  1117   515   106
#> # … with 348 more rows, and 14 more variables: to <dbl>, tpa <dbl>, tpm <dbl>,
#> #   fg_pct <dbl>, tp_pct <dbl>, ft_pct <dbl>, rpg <dbl>, apg <dbl>, spg <dbl>,
#> #   bpg <dbl>, tpg <dbl>, fpg <dbl>, ppg <dbl>, games <int>