Get Team History

Returns efficiency and four factor finish, conference and overall records, and basic shooting percentages by year back to 2008.

Usage

bart_team_history(team)

Arguments

team

Indicates team to return.

Value

Returns a tibble with 23 columns:

year

double.

t_rk

double.

coach

character.

seed

double.

finish

character.

conf

character.

ov_rec

character.

conf_rec

character.

adj_o

double.

adj_d

double.

efg_o

double.

efg_d

double.

to_percent

double.

tod_percent

double.

or_percent

double.

dr_percent

double.

ftr

double.

ftrd

double.

ft_percent

double.

two_pct

double.

two_pct_d

double.

three_pct

double.

adj_t

double.

Details

`team` input must be an exact match: 'State' is abbreviated to 'St.' and full names are used where applicable (e.g. 'North Carolina'). For complete list, see `team` column of bart_ratings.

Examples

bart_team_history(team='Charlotte')
#> # A tibble: 15 × 23
#>     year  t_rk coach    seed finish conf  ov_rec conf_…¹ adj_o adj_d efg_o efg_d
#>    <dbl> <dbl> <chr>   <dbl> <chr>  <chr> <chr>  <chr>   <dbl> <dbl> <dbl> <dbl>
#>  1  2022   191 Ron Sa…    NA NA     CUSA  17-14  10-8    105.  106.   53.0  52.6
#>  2  2021   252 Ron Sa…    NA NA     CUSA  9-16   5-11     96.1 104.   49.2  50.9
#>  3  2020   162 Ron Sa…    NA NA     CUSA  16-13  10-8     98.8  98.1  51.7  50.4
#>  4  2019   314 Ron Sa…    NA NA     CUSA  8-21   5-13     95.7 109.   48.1  51.3
#>  5  2018   315 Mark P…    NA NA     CUSA  6-23   2-16     98.6 113.   48.3  54.7
#>  6  2017   226 Mark P…    NA NA     CUSA  13-17  7-11    102.  108.   50.5  52.3
#>  7  2016   218 Mark P…    NA NA     CUSA  14-19  9-9     103.  108.   51.0  54.8
#>  8  2015   166 Ryan O…    NA NA     CUSA  14-18  7-11    106.  106.   51.8  50.6
#>  9  2014   171 Alan M…    NA NA     CUSA  17-14  7-9     102.  102.   49.7  49.2
#> 10  2013   128 Alan M…    NA NA     A10   21-12  8-8      99.8  96.2  47.1  46.9
#> 11  2012   145 Alan M…    NA NA     A10   13-17  5-11     99.2  96.9  46.9  48.9
#> 12  2011   212 Alan M…    NA NA     A10   10-20  2-14     99.5 104.   47.6  49.0
#> 13  2010   104 Bobby …    NA NA     A10   19-12  9-7     107.   99.9  47.4  47.2
#> 14  2009   137 Bobby …    NA NA     A10   11-20  5-11    102.  100.   46.9  49.6
#> 15  2008    89 Bobby …    NA NA     A10   20-14  9-7     103.   93.9  47.1  49.0
#> # … with 11 more variables: to_percent <dbl>, tod_percent <dbl>,
#> #   or_percent <dbl>, dr_percent <dbl>, ftr <dbl>, ftrd <dbl>,
#> #   ft_percent <dbl>, two_pct <dbl>, two_pct_d <dbl>, three_pct <dbl>,
#> #   adj_t <dbl>, and abbreviated variable name ¹​conf_rec