Get Team Schedule

Returns full team schedule and game-by-game four factor statistics by year

Usage

bart_team_schedule(year = current_season(), team = NULL)

Arguments

year

Defaults to current season (YYYY).

team

Team to return.

Value

Returns a tibble of game-by-game four factors by team

Details

For a brief explanation of each factor and its computation, please visit KenPom's blog. `lead_diff` is the average lead or deficit during a game for the team.

Examples

bart_team_schedule(year=2022, team='Duke')
#> # A tibble: 39 × 29
#>    date       type   team  conf  opp     opp_conf loc   result points opp_points
#>    <date>     <chr>  <chr> <chr> <chr>   <chr>    <chr> <chr>   <dbl>      <dbl>
#>  1 2022-04-02 post   Duke  ACC   North … ACC      N     L          77         81
#>  2 2022-03-26 post   Duke  ACC   Arkans… SEC      N     W          78         69
#>  3 2022-03-24 post   Duke  ACC   Texas … B12      N     W          78         73
#>  4 2022-03-20 post   Duke  ACC   Michig… B10      N     W          85         76
#>  5 2022-03-18 post   Duke  ACC   Cal St… BW       N     W          78         61
#>  6 2022-03-12 conf_t Duke  ACC   Virgin… ACC      N     L          67         82
#>  7 2022-03-11 conf_t Duke  ACC   Miami … ACC      N     W          80         76
#>  8 2022-03-10 conf_t Duke  ACC   Syracu… ACC      N     W          88         79
#>  9 2022-03-05 conf   Duke  ACC   North … ACC      H     L          81         94
#> 10 2022-03-01 conf   Duke  ACC   Pittsb… ACC      A     W          86         56
#> # … with 29 more rows, and 19 more variables: lead_diff <dbl>, adj_o <dbl>,
#> #   adj_d <dbl>, ppp <dbl>, efg <dbl>, to <dbl>, or <dbl>, ftr <dbl>,
#> #   def_ppp <dbl>, def_efg <dbl>, def_to <dbl>, def_or <dbl>, def_ftr <dbl>,
#> #   game_score <dbl>, year <dbl>, poss <dbl>, coach <chr>, opp_coach <chr>,
#> #   game_id <chr>