T-Rank Tournament Prediction

Returns single-elimination tournament predictions using Barttorvik's model

Usage

bart_tournament_prediction(teams = NULL, date = NULL, sims = NULL, seed = NULL)

Arguments

teams

List of teams arranged in scheduled order

date

Date for tournament

sims

Number of simulations to play

seed

Seed to set for reproducible results

Value

Returns a tibble with four columns:

team

character.

wins

integer.

finals

integer.

champ

integer.

Details

Given a list of teams arranged in scheduled order, function returns tournament predictions for single-elimination tournaments. Teams must be supplied as a character vector and in bracket order. In other words, if team A plays team B and team C plays team D in round one, the function should be supplied `teams=c('A', 'B', 'C', 'D')`.

Examples

bart_tournament_prediction(teams=c('Duke', 'North Carolina',
'Kansas', 'Villanova'), date='20220402', sims=10, seed=1)
#> ── Tournament Prediction: 10 Sims ────────────────────────────── toRvik 1.1.0 ──
#> ℹ Data updated: 2022-09-09 23:43:36 UTC
#> # A tibble: 4 × 4
#>   team            wins finals champ
#>   <chr>          <int>  <int> <int>
#> 1 Kansas            11      7     4
#> 2 Duke               8      5     3
#> 3 Villanova          5      3     2
#> 4 North Carolina     6      5     1