Get NCAA Tournament Results

Returns a tibble with raw and adjusted NCAA tournament results by team, coach, conference, or seed back to 2000.

Usage

bart_tourney_results(
  min_year = 2000,
  max_year = current_season(),
  type = "team"
)

Arguments

min_year

Minimum year to pull -- defaults to 2000 (YYYY).

max_year

Maximum year to pull -- defaults to current season (YYYY).

type

Data split value, defaults to team ('team', 'coach', 'conf', 'seed').

Value

Returns a tibble of adjusted and raw tournament results.

Details

PAKE is the number of wins above or below the expected value given a KenPom rating.

Examples

bart_tourney_results(min_year=2010, max_year=2015, type='conf')
#> # A tibble: 32 × 18
#>    conf   pake  pase  wins  loss w_percent   r64   r32   s16    e8    f4    f2
#>    <chr> <dbl> <dbl> <dbl> <dbl>     <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 SEC     7.8   9.6    43    24     0.642    25    14    11    10     5     2
#>  2 Horz    6.4   7.6    10     6     0.625     6     2     2     2     2     2
#>  3 P12     6     8.4    29    20     0.592    20    16    10     3     0     0
#>  4 B10     5.4   7.5    63    37     0.63     37    27    18     9     6     2
#>  5 CAA     3     3.4     7     8     0.467     8     4     1     1     1     0
#>  6 ACC     2.8   1      51    30     0.63     32    24    14     7     2     2
#>  7 Amer    2.4   4      10     5     0.667     6     4     2     1     1     1
#>  8 MVC     1.7   2.6    12     9     0.571     9     6     2     1     1     0
#>  9 Ivy     1.4   1.8     4     6     0.4       6     3     1     0     0     0
#> 10 ASun    0.8   1.7     3     6     0.333     6     2     1     0     0     0
#> # … with 22 more rows, and 6 more variables: champ <dbl>, top2 <dbl>,
#> #   f4_percent <dbl>, champ_percent <dbl>, from <dbl>, to <dbl>