Get Transfer Portal Statistics

Returns detailed, season-long player statistics on a variety of splits for all portal players for 2022-23. Transfer information is sourced from Verbal Commits and is updated every 20 minutes. Commit information is updated roughly every 18 hours.

Usage

bart_transfers(stat = NULL, conf_only = F, active = T)

Arguments

stat

Indicates statistical split (see details).

conf_only

Logical. Filters data by conference-only play; defaults to `FALSE`.

active

Logical. Filters players by portal status -- active vs. committed; defaults to TRUE (active).

Details

Data is split on three statistical types, explained below:

box

Returns basic box score stats.

shooting

Returns play-by-play shooting splits.

adv

Returns advanced metrics and possession-adjusted box score statistics.

Examples

bart_transfers(stat='box')
#> # A tibble: 961 × 20
#>    player     team  pos   exp   hgt   conf      g   mpg   ppg fg_pct  oreb  dreb
#>    <chr>      <chr> <chr> <chr> <chr> <chr> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl>
#>  1 A.J. Hamr… Char… C     Fr    6-9   BSth      5  6.75  3.38  0.167 1.38  1.75 
#>  2 A.J. McGi… Cinc… Wing… So    6-3   Amer      7  4.88  2.25  0.35  0     0.5  
#>  3 A.J. Oliv… Old … Wing… Sr    6-5   CUSA     31 17.6   2.56  0.310 0.469 1.5  
#>  4 AJ Neal    Long… Wing… Fr    6-1   BW       14  5.75  1.88  0.4   0.125 0.438
#>  5 Aaron Cla… Sacr… Scor… Sr    6-1   NEC      25 33.1  16.0   0.421 0.36  1.32 
#>  6 Abdou Ndi… Illi… PF/C  Jr    6-9   MVC      24  9.65  1.69  0.55  0.769 1.15 
#>  7 Abdoul Ka… St. … PF/C  So    6-9   A10      33  9.58  3.73  0.513 0.758 0.939
#>  8 Adam Anho… West… PF/C  Jr    6-8   Sum      11  4.46  1.31  0.455 0.462 0.923
#>  9 Adam This… Colo… Stre… Sr    6-6   MWC      23 12.5   2.29  0.38  0.25  1.5  
#> 10 Adong Mak… Nort… C     Sr    6-9   MAC      30 16.9   5.77  0.574 1.5   1.83 
#> # … with 951 more rows, and 8 more variables: rpg <dbl>, apg <dbl>,
#> #   ast_to <dbl>, spg <dbl>, bpg <dbl>, num <chr>, year <dbl>, id <dbl>