Get Team Game Stats

Returns box score data by team for all games.

Usage

bart_game_box(
  year = current_season(),
  date = NULL,
  team = NULL,
  conf = NULL,
  game_id = NULL
)

Arguments

year

Filters to year (YYYY)

date

Filters to date (YYYY-MM-DD)

team

Filters to team

conf

Filters to conference

game_id

Filters to game

Value

Returns a tibble of box score statistics

Details

The home team is coded as `team2`. Neutral site games may contain errors as to whom is the home team.

Examples

bart_game_box(year=2022)
#> ── Game Box ──────────────────────────────────────────────────── toRvik 1.1.0 ──
#> ℹ Data updated: 2022-09-09 12:23:43 UTC
#> # A tibble: 11,934 × 48
#>    date  type  locat…¹ did_win team  conf  opp   opp_c…²   min   pos   fgm   fga
#>    <chr> <chr> <chr>   <lgl>   <chr> <chr> <chr> <chr>   <int> <int> <int> <int>
#>  1 2022… post  N       FALSE   Nort… ACC   Kans… B12       200    73    23    73
#>  2 2022… post  N       TRUE    Kans… B12   Nort… ACC       200    73    29    66
#>  3 2022… post  N       FALSE   Vill… BE    Kans… B12       200    57    22    57
#>  4 2022… post  N       TRUE    Kans… B12   Vill… BE        200    57    29    54
#>  5 2022… post  N       TRUE    Nort… ACC   Duke  ACC       200    69    27    64
#>  6 2022… post  N       FALSE   Duke  ACC   Nort… ACC       200    69    30    72
#>  7 2022… post  A       TRUE    Fres… MWC   Coas… SB        200    66    31    59
#>  8 2022… post  H       FALSE   Coas… SB    Fres… MWC       200    66    25    63
#>  9 2022… post  N       TRUE    Xavi… BE    Texa… SEC       200    67    29    60
#> 10 2022… post  N       FALSE   Texa… SEC   Xavi… BE        200    67    23    51
#> # … with 11,924 more rows, 36 more variables: tpm <int>, tpa <int>, ftm <int>,
#> #   fta <int>, oreb <int>, dreb <int>, reb <int>, ast <int>, stl <int>,
#> #   blk <int>, to <int>, pf <int>, pts <int>, second_chance_pts <dbl>,
#> #   second_chance_fgm <dbl>, second_chance_fga <dbl>,
#> #   second_change_fg_pct <dbl>, pts_in_paint <dbl>, pts_in_paint_fgm <dbl>,
#> #   pts_in_paint_fga <dbl>, pts_in_paint_fg_pct <dbl>, fast_brk_pts <dbl>,
#> #   fast_brk_fgm <dbl>, fast_brk_fga <dbl>, fast_brk_fg_pct <dbl>, …